Add message history, console-style
With a heartbeat sent from the experiment, useful messages are
over-written in the display window too quickly. Keeping the raw history
is also hard to read due to repeat heartbeat messages. This implements
history of message display with a message repetition count
instead. There seems to be an issue with long messages getting double
counted.

[Issue: 2]